chromadb-central/requirements.txt
fastapi
uvicorn
chromadb
python-dotenv
google-auth
google-api-python-client
google-cloud-storage
openai
PyPDF2
python-docx
tqdm
